fix: 자동로그인 수정(이름 변경 및 key값 수정)
